# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Marketplace to connect students and teachers",
        "title": "Study marketplace API",
        "termsOfService": "[TODO]",
        "contact": {
            "name": "API Support",
            "url": "[TODO]",
            "email": "[TODO]"
        },
        "license": {
            "name": "[TODO]",
            "url": "[TODO]"
        },
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/api/": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Show the status of server.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "requires email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "POST request for login",
                "parameters": [
                    {
                        "description": "request info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/auth/password-reset": {
            "post": {
                "description": "requires registred email address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "password-reset"
                ],
                "summary": "POST request to update password",
                "parameters": [
                    {
                        "description": "user email for update",
                        "name": "password-reset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "requires username and password for registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "register"
                ],
                "summary": "POST request for registration",
                "parameters": [
                    {
                        "description": "user info for sign in",
                        "name": "user_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/queries.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protected/advertisement-create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "endpoint for advertisement creation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisement-create"
                ],
                "summary": "POST request to create advertisement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "advertisement information",
                        "name": "advertisement-create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/queries.Advertisement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protected/advertisement-delete": {
            "delete": {
                "description": "endpoint for advertisement deletion by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisement-delete"
                ],
                "summary": "PATCH request to delete advertisement",
                "parameters": [
                    {
                        "description": "advertisement id",
                        "name": "advertisement-delete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protected/advertisement-patch": {
            "patch": {
                "description": "endpoint for advertisement update",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisement-patch"
                ],
                "summary": "PATCH request to update advertisement",
                "parameters": [
                    {
                        "description": "advertisement information",
                        "name": "advertisement-patch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/queries.Advertisement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protected/user-patch": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "requires valid token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-patch"
                ],
                "summary": "PATCH request to update user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "user info for update",
                        "name": "userinfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/queries.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protected/userinfo": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "requires valid token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userinfo"
                ],
                "summary": "Get request to see user info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.EmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.Id": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.InLogin": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "queries.Advertisement": {
            "type": "object",
            "properties": {
                "attachment": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "mobile_phone": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "telegram": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "queries.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}